ingestion.py
from dotenv import load_dotenv
from langchain_chroma import Chroma
from langchain_community.document_loaders import WebBaseLoader
from langchain_openai import OpenAIEmbeddings
from langchain_text_splitters import RecursiveCharacterTextSplitter

load_dotenv()

urls = [
    "https://lilianweng.github.io/posts/2023-06-23-agent/",
    "https://lilianweng.github.io/posts/2023-03-15-prompt-engineering/",
    "https://lilianweng.github.io/posts/2023-10-25-adv-attack-llm/",
]

docs = [WebBaseLoader(url).load() for url in urls]
docs_list = [item for sublist in docs for item in sublist]

text_splitter = RecursiveCharacterTextSplitter(chunk_size=250, chunk_overlap=0)

doc_splits = text_splitter.split_documents(docs_list)

# vector_store = Chroma.from_documents(
#     documents=doc_splits,
#     persist_directory="./.chroma_db",
#     embedding=OpenAIEmbeddings(),
#     collection_name="rag-data"
# )


retriever = Chroma(
    collection_name="rag-data",
    persist_directory="./.chroma_db",
    embedding_function=OpenAIEmbeddings(),
)
