pyproject.toml
[project]
name = "rag-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "black>=26.3.1",
    "grandalf>=0.8",
    "isort>=8.0.1",
    "langchain>=1.2.14",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.12",
    "langchain-tavily>=0.2.17",
    "langchainhub>=0.1.21",
    "langgraph>=1.1.4",
    "langgraph-cli[inmem]>=0.4.19",
    "langsmith>=0.7.23",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
]
